Speed 0 error
Rychlost byla zaokrouhlována dolů na 0, pokud byla vybrána rychlost
menší než 1

+++ public_html/js/view/ctrl_view.js
@@ -88,7 +88,7 @@ $.extend(view.ctrl,{
                 rendered=Mustache.render(template,{poss:obj});
                 //$(rendered).val(val);
                 this.summonSelect(event.currentTarget,rendered,function(event){
-                    sett[ctrl].set(parseInt(event.target.getAttribute("data-val"),10));
+                    sett[ctrl].set(parseFloat(event.target.getAttribute("data-val"),10));
                     $("#ctrl_select").hide();
                 });
                 return;
